+++ src/routes/v2/projects/[projectId]/workflows/[workflowId]/+page.server.js
@@ -2,7 +2,6 @@ import { getWorkflow, getWorkflowJobs } from '$lib/server/api/v2/workflow_api';
 import { getProjectDatasets } from '$lib/server/api/v2/project_api';
 import { getDefaultWorkflowDataset } from '$lib/common/workflow_utilities';
 import { getLogger } from '$lib/server/logger.js';
-import { getCurrentUser } from '$lib/server/api/auth_api';
 
 const logger = getLogger('workflow page [v2]');
 
@@ -11,10 +10,6 @@ export async function load({ fetch, params }) {
 
 	const { projectId, workflowId } = params;
 
-	const user =
-		/** @type {import('fractal-components/types/api').User & {group_ids_names: Array<[number, string]>}} */ (
-			await getCurrentUser(fetch, true)
-		);
 	const workflow = await getWorkflow(fetch, projectId, workflowId);
 	const datasets = await getProjectDatasets(fetch, projectId);
 	const jobs = await getWorkflowJobs(fetch, projectId, workflowId);
@@ -24,7 +19,6 @@ export async function load({ fetch, params }) {
 	return {
 		workflow,
 		datasets,
-		defaultDatasetId,
-		user
+		defaultDatasetId
 	};
 }