+++ src/routes/v2/tasks/management/+page.server.js
@@ -1,4 +1,3 @@
-import { removeIdenticalTaskGroups } from 'fractal-components/tasks/task_group_utilities';
 import { getCurrentUser } from '$lib/server/api/auth_api.js';
 import { listTaskGroups } from '$lib/server/api/v2/task_api';
 
@@ -7,7 +6,7 @@ export async function load({ fetch }) {
 		/** @type {import('fractal-components/types/api').User & {group_ids_names: Array<[number, string]>}} */ (
 			await getCurrentUser(fetch, true)
 		);
-	const taskGroups = removeIdenticalTaskGroups(await listTaskGroups(fetch), user);
+	const taskGroups = await listTaskGroups(fetch);
 
 	return {
 		user,